cannot pass 2833 on a transcoded call
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@8386 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_rtp.h
@@ -233,7 +233,7 @@ SWITCH_DECLARE(void) switch_rtp_set_flag(switch_rtp_t *rtp_session, switch_rtp_f
   \param flags the flags to test
   \return TRUE or FALSE
 */
-SWITCH_DECLARE(uint8_t) switch_rtp_test_flag(switch_rtp_t *rtp_session, switch_rtp_flag_t flags);
+SWITCH_DECLARE(uint32_t) switch_rtp_test_flag(switch_rtp_t *rtp_session, switch_rtp_flag_t flags);
 
 /*!
   \brief Clear an RTP Flag
